% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testphrases}
\alias{testphrases}
\title{sample short documents for testing}
\description{
Some sample short documents in plain text format for testing
  with \code{\link{liwcalike}}.
}
\examples{
liwcalike(testphrases)

}

